Aula 07-03-2022

+++ 10-Poo/02-Trabalhando-com-Arquivos/Helper/FileHelper.cs
@@ -26,8 +26,86 @@ namespace Trabalhando_com_Arquivos.Helper
         }
 
         public void DeletarDiretorios(string caminho) {
-            Directory.Delete(caminho);         
+            if (Directory.Exists(caminho)) 
+            {
+                Directory.Delete(caminho);         
+            }
         }
 
+        public void CriarArquivo(string caminho, string conteudo) {
+            if (!File.Exists(caminho)) 
+            {
+                File.WriteAllText(caminho, conteudo);
+            }            
+        }
+
+        public void CriarArquivoTextoStream(string caminho, List<string> conteudo)  {
+            using (var stream = File.CreateText(caminho))
+            {
+                 foreach (var linha in conteudo)
+                {
+                    stream.WriteLine(linha);
+                }                    
+            }
+        }    
+
+        public void AdicionarTexto(string caminho, string conteudoAdicionar)
+        {
+            File.AppendAllText(caminho, conteudoAdicionar);
+        }
+
+        public void AdicionarTextoStream(string caminho, List<string> conteudoAdicionar)
+        {
+            using (var stream = File.AppendText(caminho))
+            {
+                 foreach (var linha in conteudoAdicionar)
+                {
+                    stream.WriteLine(linha);
+                }                    
+            }
+        }
+    
+        public void LerAquivo(string caminho) {
+            var conteudo = File.ReadAllLines(caminho);
+            
+            foreach (var linha in conteudo)
+            {
+                System.Console.WriteLine(linha);        
+            }
+        }
+
+        public void LerAquivoStream(string caminho) {
+
+            string linha = string.Empty;
+
+            using (var conteudoStream = File.OpenText(caminho))
+            {
+                while ((linha = conteudoStream.ReadLine())
+                != null)
+                {
+                    System.Console.WriteLine(linha);
+                }
+            }
+        }
+
+        public void MoverArquivo(string caminho, string novoCaminho, bool sobrescrever){
+            if (sobrescrever) {
+                File.Move(caminho, novoCaminho);
+            }
+        }
+        
+        public void CopiarArquivo(string caminho, string novoCaminho, bool sobrescrever){
+            if (sobrescrever) {
+                File.Copy(caminho, novoCaminho, sobrescrever);
+            }            
+        }
+
+        public void DeletarArquivo(string caminho, string nome) {
+
+            if ((File.Exists(caminho)) && Path.GetFileName(caminho).Equals(nome)) {
+                File.Delete(caminho);
+            }
+        }
+        
     }
 }
